+++ src/ecs/task.ts
@@ -3,6 +3,7 @@ import { RunTaskCommandInput } from '@aws-sdk/client-ecs'
 import { Configuration, ConfiguredVariables } from '../types/configuration'
 
 interface Params {
+  clusterName: string
   task: string
   revision: string
   variables: ConfiguredVariables
@@ -10,15 +11,15 @@ interface Params {
 }
 
 export const taskFromConfiguration = (params: Params): RunTaskCommandInput => {
-  const { task, revision, variables, config } = params
+  const { clusterName, task, revision, variables, config } = params
   const { project, environment } = variables
 
-  const clusterConfig = config.clusters[environment]
+  const clusterConfig = config.clusters[clusterName]
   const subnets = clusterConfig.publicSubnets
   const securityGroups = clusterConfig.securityGroups
 
   return {
-    cluster: `${project}-${environment}`,
+    cluster: clusterName,
     taskDefinition: `${project}-${task}-${environment}:${revision}`,
     count: 1,
     group: `task:${task}`,